User and Role tests separated
